Renamed test files.
In Django 1.6, the default test runner was changed to DiscoverRunner which
ignores files whose filenames do not start with “test”.

The imports in __init__.py files are no longer needed in Django 1.6, but I’ve
preserved them to maintain compatibility with previous versions.
