+++ tests/tests/__init__.py
@@ -3,5 +3,5 @@ from integrations import *
 from handlers import *
 from models import *
 from views import *
-from .decorators import TestDecoratorErrors
-from .middleware import TestMiddleware
+from .test_decorators import TestDecoratorErrors
+from .test_middleware import TestMiddleware
