+++ tests/tests/handlers/__init__.py
@@ -1,1 +1,1 @@
-from .bearer import TestBearerHandler
+from .test_bearer import TestBearerHandler
