+++ tests/tests/handlers/test_bearer.py
