+++ tests/tests/integrations/__init__.py
@@ -1,2 +1,2 @@
-from .idan_oauthlib import TestOauthlib
-from .litl_rauth import TestRauth
+from .test_idan_oauthlib import TestOauthlib
+from .test_litl_rauth import TestRauth
