+++ tests/tests/models/__init__.py
@@ -1,3 +1,3 @@
-from .authorization_token import TestAuthorizationTokenModel
-from .client import TestClientModel
-from .redirect_uri import TestRedirectUriModel
+from .test_authorization_token import TestAuthorizationTokenModel
+from .test_client import TestClientModel
+from .test_redirect_uri import TestRedirectUriModel
