+++ tests/tests/models/test_authorization_token.py
